requirements.txt
openai>=1.0.0
google-genai>=0.2.0
google-generativeai>=0.7.2
feedparser
yagmail
pyyaml
beautifulsoup4
requests
lxml